[Templates] Update pipeline extension template
