+++ ProjectTemplates/DotNetTemplate/MonoGame.Templates.CSharp/content/MonoGame.Library.Pipeline/.template.config/template.json
@@ -1,0 +1,29 @@
+{
+    "author": "MonoGame Team",
+    "classifications": [
+        "MonoGame"
+    ],
+    "name": "MonoGame Pipeline Extension",
+    "groupIdentity": "MonoGame.Library.Pipeline",
+    "identity": "MonoGame.Library.Pipeline.CSharp",
+    "shortName": "mgpipeline",
+    "tags": {
+        "language": "C#",
+        "type": "project"
+    },
+    "primaryOutputs": [
+        {
+            "path": "MGNamespace.csproj"
+        }
+    ],
+    "sourceName": "MGNamespace",
+    "preferNameDirectory": "true",
+    "description": "A MonoGame content pipeline extension library project.",
+    "symbols": {
+        "framework": {
+            "type": "parameter",
+            "defaultValue": "net45",
+            "replaces": "netstandard2.0"
+        }
+    }
+}
