+++ ProjectTemplates/DotNetTemplate/MonoGame.Templates.CSharp/content/MonoGame.Library.Pipeline/MGNamespace.csproj
@@ -1,0 +1,16 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>netstandard2.0</TargetFramework>
+  </PropertyGroup>
+
+  <ItemGroup>
+     <PackageReference Include="MonoGame.Framework.Portable" Version="3.7.0.1708">
+      <PrivateAssets>All</PrivateAssets>
+    </PackageReference>
+     <PackageReference Include="MonoGame.Framework.Content.Pipeline.Portable" Version="3.7.0.1708">
+      <PrivateAssets>All</PrivateAssets>
+    </PackageReference>
+  </ItemGroup>
+
+</Project>
