+++ ProjectTemplates/DotNetTemplate/MonoGame.Templates.CSharp/content/MonoGame.Library.Pipeline/Processor1.cs
@@ -1,0 +1,16 @@
+﻿using Microsoft.Xna.Framework.Content.Pipeline;
+
+using TInput = System.String;
+using TOutput = System.String;
+
+namespace MGNamespace
+{
+    [ContentProcessor(DisplayName = "Processor1")]
+    class Processor1 : ContentProcessor<TInput, TOutput>
+    {
+        public override TOutput Process(TInput input, ContentProcessorContext context)
+        {
+            return default(TOutput);
+        }
+    }
+}
