--- ProjectTemplates/DotNetTemplate/MonoGame.Templates.CSharp/content/PipelineExtension/.template.config/dotnetcli.host.json
@@ -1,3 +1,0 @@
-{
-
-}
