--- ProjectTemplates/DotNetTemplate/MonoGame.Templates.CSharp/content/PipelineExtension/Processor1.cs
@@ -1,23 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using Microsoft.Xna.Framework;
-using Microsoft.Xna.Framework.Graphics;
-using Microsoft.Xna.Framework.Content.Pipeline;
-using Microsoft.Xna.Framework.Content.Pipeline.Graphics;
-
-using TInput = System.String;
-using TOutput = System.String;
-
-namespace BlankApp
-{
-	[ContentProcessor(DisplayName = "Processor1")]
-	class Processor1 : ContentProcessor<TInput, TOutput>
-	{
-		public override TOutput Process(TInput input, ContentProcessorContext context)
-		{
-			return default(TOutput);
-		}
-	}
-}
-
