--- ProjectTemplates/DotNetTemplate/MonoGame.Templates.CSharp/content/PipelineExtension/packages.config
@@ -1,5 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="MonoGame.Framework.Content.Pipeline.Portable" version="3.2.99.1-Beta" targetFramework="net45" />
-  <package id="MonoGame.Framework.Portable" version="3.2.99.1-Beta" targetFramework="net45" />
-</packages>
